* Create a processor "type" attribute that can be used to map each processor type to different chains of handlers * Add frontend type endpoints. The processing pipeline is the same as for backend endpoints for now, but they might require different handlers * Handle /healthcheck like any other processor
