+++ beater/server_test.go
@@ -2,9 +2,7 @@ package beater
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/http/httptest"
@@ -18,6 +16,7 @@ import (
 	"github.com/kabukky/httpscerts"
 	"github.com/stretchr/testify/assert"
 
+	"github.com/elastic/apm-server/processor/healthcheck"
 	"github.com/elastic/apm-server/processor/transaction"
 	"github.com/elastic/apm-server/tests"
 	"github.com/elastic/beats/libbeat/beat"
@@ -41,24 +40,6 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-func TestDecode(t *testing.T) {
-	transactionBytes, err := tests.LoadValidData("transaction")
-	assert.Nil(t, err)
-	buffer := bytes.NewReader(transactionBytes)
-
-	req, err := http.NewRequest("POST", "_", buffer)
-	req.Header.Add("Content-Type", "application/json")
-	assert.Nil(t, err)
-
-	res, err := decodeData(req)
-	assert.Nil(t, err)
-	assert.NotNil(t, res)
-
-	body, err := ioutil.ReadAll(res)
-	assert.Nil(t, err)
-	assert.Equal(t, transactionBytes, body)
-}
-
 func TestServerOk(t *testing.T) {
 	apm, teardown := setupServer(t, noSSL)
 	defer teardown()
@@ -69,6 +50,20 @@ func TestServerOk(t *testing.T) {
 	rr := httptest.NewRecorder()
 	apm.Handler.ServeHTTP(rr, req)
 	assert.Equal(t, 202, rr.Code, rr.Body.String())
+}
+
+func TestServerHealth(t *testing.T) {
+	apm, teardown := setupServer(t, noSSL)
+	defer teardown()
+
+	req, err := http.NewRequest("GET", healthcheck.Endpoint, nil)
+	if err != nil {
+		t.Fatalf("Failed to create test request object: %v", err)
+	}
+
+	rr := httptest.NewRecorder()
+	apm.Handler.ServeHTTP(rr, req)
+	assert.Equal(t, 200, rr.Code, rr.Code)
 }
 
 func TestServerNoContentType(t *testing.T) {
@@ -132,68 +127,6 @@ func TestServerBadProtocol(t *testing.T) {
 	assert.Contains(t, err.Error(), "malformed HTTP response")
 }
 
-func TestJSONFailureResponse(t *testing.T) {
-	req, err := http.NewRequest("POST", "/transactions", nil)
-	assert.Nil(t, err)
-
-	req.Header.Set("Accept", "application/json")
-	w := httptest.NewRecorder()
-
-	sendStatus(w, req, 400, errors.New("Cannot compare apples to oranges"))
-
-	resp := w.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
-	assert.Equal(t, 400, w.Code)
-	assert.Equal(t, body, []byte(`{"error":"Cannot compare apples to oranges"}`))
-	assert.Equal(t, "application/json", resp.Header.Get("Content-Type"))
-}
-
-func TestJSONFailureResponseWhenAcceptingAnything(t *testing.T) {
-	req, err := http.NewRequest("POST", "/transactions", nil)
-	assert.Nil(t, err)
-	req.Header.Set("Accept", "*/*")
-	w := httptest.NewRecorder()
-
-	sendStatus(w, req, 400, errors.New("Cannot compare apples to oranges"))
-
-	resp := w.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
-	assert.Equal(t, 400, w.Code)
-	assert.Equal(t, body, []byte(`{"error":"Cannot compare apples to oranges"}`))
-	assert.Equal(t, "application/json", resp.Header.Get("Content-Type"))
-}
-
-func TestHTMLFailureResponse(t *testing.T) {
-	req, err := http.NewRequest("POST", "/transactions", nil)
-	assert.Nil(t, err)
-	req.Header.Set("Accept", "text/html")
-	w := httptest.NewRecorder()
-
-	sendStatus(w, req, 400, errors.New("Cannot compare apples to oranges"))
-
-	resp := w.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
-	assert.Equal(t, 400, w.Code)
-	assert.Equal(t, body, []byte(`Cannot compare apples to oranges`))
-	assert.Equal(t, "text/plain; charset=utf-8", resp.Header.Get("Content-Type"))
-}
-
-func TestFailureResponseNoAcceptHeader(t *testing.T) {
-	req, err := http.NewRequest("POST", "/transactions", nil)
-	assert.Nil(t, err)
-
-	req.Header.Del("Accept")
-
-	w := httptest.NewRecorder()
-	sendStatus(w, req, 400, errors.New("Cannot compare apples to oranges"))
-
-	resp := w.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
-	assert.Equal(t, 400, w.Code)
-	assert.Equal(t, body, []byte(`Cannot compare apples to oranges`))
-	assert.Equal(t, "text/plain; charset=utf-8", resp.Header.Get("Content-Type"))
-}
-
 func setupServer(t *testing.T, ssl *SSLConfig) (*http.Server, func()) {
 	if testing.Short() {
 		t.Skip("skipping server test")
@@ -233,7 +166,7 @@ func withSSL(t *testing.T, domain string) *SSLConfig {
 }
 
 func makeTestRequest(t *testing.T) *http.Request {
-	req, err := http.NewRequest("POST", transaction.Endpoint, bytes.NewReader(testData))
+	req, err := http.NewRequest("POST", transaction.BackendEndpoint, bytes.NewReader(testData))
 	if err != nil {
 		t.Fatalf("Failed to create test request object: %v", err)
 	}
@@ -246,7 +179,7 @@ func postTestRequest(t *testing.T, apm *http.Server, client *http.Client, schema
 		client = http.DefaultClient
 	}
 
-	addr := fmt.Sprintf("%s://%s%s", schema, apm.Addr, transaction.Endpoint)
+	addr := fmt.Sprintf("%s://%s%s", schema, apm.Addr, transaction.BackendEndpoint)
 	return client.Post(addr, "application/json", bytes.NewReader(testData))
 }
