+++ include/list.go
@@ -8,5 +8,6 @@ package include
 import (
 	// This list is automatically generated by `make imports`
 	_ "github.com/elastic/apm-server/processor/error"
+	_ "github.com/elastic/apm-server/processor/healthcheck"
 	_ "github.com/elastic/apm-server/processor/transaction"
 )
