+++ processor/error/benchmark_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func BenchmarkEventWithFileLoading(b *testing.B) {
-	processor := NewProcessor()
+	processor := NewBackendProcessor()
 	for i := 0; i < b.N; i++ {
 		data, _ := tests.LoadValidData("error")
 		err := processor.Validate(data)
@@ -20,7 +20,7 @@ func BenchmarkEventWithFileLoading(b *testing.B) {
 }
 
 func BenchmarkEventFileLoadingOnce(b *testing.B) {
-	processor := NewProcessor()
+	processor := NewBackendProcessor()
 	data, _ := tests.LoadValidData("error")
 	for i := 0; i < b.N; i++ {
 		err := processor.Validate(data)
