+++ processor/error/package_tests/fields_test.go
@@ -14,7 +14,7 @@ func TestFields(t *testing.T) {
 		"./../../../_meta/fields.common.yml",
 		"./../_meta/fields.yml",
 	}
-	tests.TestEventAttrsDocumentedInFields(t, fieldsPaths, er.NewProcessor)
+	tests.TestEventAttrsDocumentedInFields(t, fieldsPaths, er.NewBackendProcessor)
 
 	notInEvent := set.New(
 		"context.db.instance",
@@ -26,5 +26,5 @@ func TestFields(t *testing.T) {
 		"error id icon",
 		"view errors",
 	)
-	tests.TestDocumentedFieldsInEvent(t, fieldsPaths, er.NewProcessor, notInEvent)
+	tests.TestDocumentedFieldsInEvent(t, fieldsPaths, er.NewBackendProcessor, notInEvent)
 }
